requirements.txt
discord.py @ git+https://github.com/rapptz/discord.py
git+https://github.com/rapptz/discord-ext-menus
git+https://github.com/gorialis/jishaku
cachetools
fuzzywuzzy
parsedatetime
python-dateutil
humanize
asyncpg
asyncpg-stubs
typing_extensions
psutil
tabulate
mystbin.py
topggpy
pillow
importlib_metadata
configparser
gitpython
python-dotenv